requirements.txt
click>=8.1
sqlalchemy>=2.0
openai>=1.30
anthropic>=0.25
pdfplumber>=0.10
pillow>=10.0
python-dotenv>=1.0
rich>=13.0
jinja2>=3.1
weasyprint>=60.0
reportlab>=4.0
pydantic>=2.0
